Refactor multi-resolution structures
